requirements.txt
arrow==0.8.0
boto3==1.4.1
botocore==1.4.71
docutils==0.12
ecdsa==0.13
Fabric==1.12.0
futures==3.0.5
jmespath==0.9.0
paramiko==1.17.2
pycrypto==2.6.1
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.9
six==1.10.0
